+++ public/index.html
@@ -8,9 +8,9 @@
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.8.2/css/all.min.css"
     />
+    <script def src="//developers.kakao.com/sdk/js/kakao.min.js"></script>
   </head>
   <body>
     <div id="root"></div>
-    <script src="//developers.kakao.com/sdk/js/kakao.min.js"></script>
   </body>
 </html>
