+++ src/pages/Main/Tealist/Detail/Detail.jsx
@@ -41,6 +41,7 @@ class Detail extends React.Component {
   };
 
   componentDidMount() {
+    this.kakaoSharePreference();
     fetch(`${GET_PRODUCT_API}/${this.props.match.params.id}`)
       .then(response => response.json())
       .then(data =>
@@ -49,7 +50,6 @@ class Detail extends React.Component {
           console.log(data.product_info[0])
         )
       );
-    // this.kakaoSharePreference();
   }
 
   includeBagPrice = id => {
@@ -79,7 +79,7 @@ class Detail extends React.Component {
               <span className="icon">
                 <i className="fas fa-shopping-bag" />
               </span>
-              <p className="content">쇼핑백동봉선택가능</p>
+              <p className="content">쇼핑백 동봉여부</p>
             </div>
           </section>
           <section className="item-info">
@@ -126,7 +126,7 @@ class Detail extends React.Component {
                 <p className="value">
                   <strong className="bold">
                     {(
-                      20000 * this.state.count +
+                      price * this.state.count +
                       this.state.bagPrice
                     ).toLocaleString()}
                   </strong>
