+++ src/pages/Main/Tealist/Detail/Sharebutton/Sharebutton.scss
@@ -19,9 +19,8 @@
     }
 
     input {
-      visibility: hidden;
       position: absolute;
-      top: 0;
+      top: -100px;
     }
   }
