+++ src/pages/Main/Tealist/FilterButton/FiterButton.scss
@@ -1,0 +1,34 @@
+@import '../../../../styles/variable.scss';
+@import '../../../../styles/variableListDetail.scss';
+@import '../../../../styles/mixinSet.scss';
+
+.filter-button {
+  display: flex;
+
+  .link {
+    flex-wrap: nowrap;
+    box-sizing: content-box;
+    min-width: 50px;
+    height: 38px;
+    margin-left: calc(#{$margin-value} - 5px);
+    padding: 0 calc(#{$margin-value} - 4px);
+    line-height: 38px;
+    font-size: $font-large;
+    text-align: center;
+    text-decoration: none;
+    color: $font-color-filterBtn;
+    background-color: $color-filterBtn;
+    border-radius: 38px;
+
+    &.active {
+      color: white;
+      background-color: $color-green;
+    }
+
+    &:hover {
+      cursor: pointer;
+      color: white;
+      background-color: $color-green;
+    }
+  }
+}
