+++ src/pages/Main/Tealist/VideoSlide/Slide/Slide.jsx
