+++ src/pages/Main/Tealist/VideoSlide/Slide/Slide.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/variable.scss';
-@import '../../../../styles/variableListDetail.scss';
+@import '../../../../../styles/variable.scss';
+@import '../../../../../styles/variableListDetail.scss';
 
 .slide {
   display: inline-block;
