+++ src/pages/Main/Tealist/VideoSlide/VideoSlide.scss
@@ -1,0 +1,89 @@
+@import '../../../../styles/variable.scss';
+@import '../../../../styles/variableListDetail.scss';
+@import '../../../../styles/mixinSet.scss';
+
+.tea-carousel {
+  @include flex-common(40%);
+  position: relative;
+  flex-direction: column;
+
+  .swiper-container {
+    width: 100%;
+
+    .swiper {
+      display: flex;
+      align-items: stretch;
+      overflow: scroll;
+      background-color: black;
+
+      .swiper-inner {
+        position: relative;
+        display: flex;
+        width: 100%;
+        height: 100%;
+        transform: translate3d(25%, 0, 0);
+        z-index: 1;
+      }
+    }
+
+    .transparentbox {
+      @include carousel-side-box;
+      z-index: 4;
+
+      &.left-0 {
+        left: 0;
+      }
+
+      &.right-0 {
+        right: 0;
+        opacity: 0.5;
+      }
+
+      .left {
+        @include position-set(absolute, 45%, 10%);
+        @include leftright-design;
+      }
+
+      .right {
+        @include position-set(absolute, 45%, auto, auto, 10%);
+        @include leftright-design;
+      }
+    }
+  }
+
+  .swiper::-webkit-scrollbar {
+    display: none;
+  }
+
+  .swiper-teaname {
+    @include flex-common(25%, center, center);
+    width: 100%;
+    min-width: 745px;
+    padding: 0 280px;
+    background-color: $bgColor-videoMenu;
+    z-index: 4;
+
+    .teaname-overflow {
+      display: flex;
+      width: 100%;
+      min-width: 745px;
+      overflow-x: scroll;
+
+      .teaname {
+        @include flex-common(40%, center);
+        padding: 40px 0;
+        text-decoration: none;
+        color: $font-color-videoMenu;
+        font-size: $font-xlarge;
+
+        &:active {
+          color: $font-color-videoMenu;
+        }
+
+        &:visited {
+          color: $font-color-videoMenu;
+        }
+      }
+    }
+  }
+}
